Hemant

+++ frontend/src/app/admincontroller/servicefiletypes/servicefiletypes.component.ts
@@ -21,6 +21,15 @@ export class ServicefiletypesComponent implements OnInit, CanServiceFileTypesCom
   filetypes = [];
   changesSaved = true;
   successMsgDiv = 'hide';
+  help = {
+    title: 'Learn How to Map Filetypes to Services',
+    instructions: [
+      "First choose any service from services listed in the dropdown.",
+      "After that you can choose filetype from the filetypes listed in the dropdown.",
+      "If the filetype is already exists in the service, software will show an error otherwise you can see the selected filetype added to the selected services filetype list.",
+      "After that you have to click on <b>Save</b> to make changes parmanent otherwise changes won't affect in the database."
+    ]
+  }
   constructor(private fileService: FileService) { }
 
   ngOnInit() {
@@ -111,15 +120,18 @@ export class ServicefiletypesComponent implements OnInit, CanServiceFileTypesCom
     );
 
     if (this.deleteFileTypeArr.length !== 0) {
-      this.fileService.deleteServiceFiletypes(this.deleteFileTypeArr).subscribe(
-        (response: Response) => {
-          console.log(response)
-          if (response.status == 200) {
-            this.changesSaved = true;
-          }
-        },
-        (error) => console.log(error)
-      );
+      let cnfrm = confirm(MESSAGE_CONST.DELETE_AUTH_CHECK);
+      if (cnfrm == true) {
+        this.fileService.deleteServiceFiletypes(this.deleteFileTypeArr).subscribe(
+          (response: Response) => {
+            console.log(response)
+            if (response.status == 200) {
+              this.changesSaved = true;
+            }
+          },
+          (error) => console.log(error)
+        );
+      }
     }
   }
